requirements.txt
# Core dependencies
openai>=1.0.0
pyyaml>=6.0

# TTS providers
edge-tts>=6.0.0
pyttsx3>=2.90

# Audio processing (for advanced features)
librosa>=0.10.0
fastdtw>=0.3.4

# Visualization
matplotlib>=3.7.0
numpy>=1.24.0

# Optional: faster WER computation
jiwer>=3.0.0
